- Resolved conflict in packages/matter-server/package.json: kept ble-proxy
  dependency while taking new matter.js version from main
- Updated packages/ble-proxy/package.json matter.js deps to match

+++ packages/custom-clusters/src/clusters/eve.ts
@@ -56,7 +56,7 @@ export class EveCluster {
     @attribute(0x130a0013, single)
     altitude?: number;
 
-    @attribute(0x130a0014, single)
+    @attribute(0x130a0014, uint32)
     pressure?: number;
 
     @attribute(0x130a0015, uint32)
